+++ src/server/server.js
@@ -9,6 +9,7 @@ const router = require('./router');
 const adminRouter = require('./routers/admin');
 const appRouter = require('./routers/app');
 const locationsRouter = require('./routers/locations');
+const lockersRouter = require('./routers/lockers');
 const config = require('./config/server.json')[env];
 const { globalErrorHandler, NotFoundError } = require('./errors');
 
@@ -48,6 +49,7 @@ app.use(session({
 app.use('/api/v1/app', appRouter);
 app.use('/api/v1/admin', adminRouter);
 app.use('/api/v1/locations', locationsRouter);
+app.use('/api/v1/lockers', lockersRouter);
 app.use('/api/v1', router);
 app.use('/api/v1/*', () => {
 	throw new NotFoundError();
